+++ include/kitty/bit_operations.hpp
@@ -33,6 +33,7 @@
 #pragma once
 
 #include <cstdint>
+#include <numeric>
 
 #include "static_truth_table.hpp"
 
@@ -115,6 +116,27 @@ template<int NumVars>
 void clear( static_truth_table<NumVars, true>& tt )
 {
   tt._bits = 0;
+}
+/*! \endcond */
+
+/*! \brief Count ones in truth table
+
+  \param tt Truth table
+*/
+template<typename TT>
+inline uint64_t count_ones( const TT& tt )
+{
+  return std::accumulate( tt.cbegin(), tt.cend(), uint64_t( 0 ),
+                          []( auto accu, auto word ) {
+                            return accu + __builtin_popcount( word & 0xffffffff ) + __builtin_popcount( word >> 32 );
+                          } );
+}
+
+/*! \cond PRIVATE */
+template<int NumVars>
+inline uint64_t count_ones( const static_truth_table<NumVars, true>& tt )
+{
+  return __builtin_popcount( tt._bits );
 }
 /*! \endcond */